depend: remove optional TOC argument to _make_toc() helper
There is exactly one call that passes the existing TOC in our
codebase, and even there we can simply append the output to the
existing TOC.

The returned toc is now a list as opposed to instance of the TOC
class. This removes implict deduplication functionality, but
should not be an issue at this step, because entries are generated
from the module graph (and so there should not be any duplicates
here).