building: work around the issue with passing code cache around
When `Analysis.pure` TOC list was an instance of the `TOC` class,
we could dynamically add an attribute `_code_cache` to it, and use
that to pass the code cache dictionary around. The `PYZ` writer
would then check for the presence of the attribute on the received
arguments and extend its internal code cache accordingly.

Now that the TOC list is a plain `list`, the custom attribute
cannot be added anymore. To work around this, we store new
dictionary in global `CONF['code_cache']`, where we associate
the list's `id()` with its code cache dictionary. This gives us
the same semantics as the old approach.