building: splash: fix detection of tkinter usage
The `Splash` target is trying to detect whether user's code already
uses `_tkinter` by looking for extension in the `binaries` TOC.
However, it uses the `filenames` property of the `TOC` class, which
is not available anymore now that the TOC lists have been switched
to plain `list`.

Futhermore, as it is trying to look up `_tkinter` extension by
module name, this means that the detection has been effectively
broken since #7273 where we pushed conversion of extension module
names to full extension filenames into `Analysis`. So the search
in `binaries` needs to be done with full extension filename rather
than module name.