tests: add tests for symbolic link collection
The tests verify that when collecting a symbolic link and its
target file, the link is collected/re-created as symbolic link
instead of a hard copy. On the other hand, when collecting only
a symbolic link without its corresponding target, the symbolic
link is collected as hard copy.

The tests cover three scenarios: symbolic link points to a
file in the same directory, to a file in sub-directory, and to
a file in a parent directory. Similarly, they cover both explicit
collection (symbolic link and corresponding file are added via
individual `--add-data switches`) and implicit collection (whole
data directory is added via `--add-data` switch).