building: implement symbolic link support
Implement basic symlink support in `Analysis` and in `PKG` and
`COLLECT` build targets.

Towards the end of `Analysis`, we now scan the combined (and
normalized) `binaries` and `datas` TOC for symbolic links. If
the original file is also being collected, we convert those
entries to SYMLINK entries, else we leave them as is (which
creates a hard copy). The SYMLINK entries are returned via
the `datas` TOC, even if their original entry came from the
`binaries` TOC.

The `PKG` build target writes the SYMLINK to the `CArchive`,
while the `COLLECT` target re-creates the symlink in the
frozen application directory.