hookutils: qt: clean up QtWebEngine file collection
Remove the explicit .framework collection (which resulted in
duplication of data). This is now unnecessary, because we preserve
original library location and thus .framework directory structure.

So all we really need is properly collect Helper and Resources
directories from QtWebEngineCore.framework.

This fixes issues with QtWebEngine in onefile builds on macOS, as
well as onedir applications being incorrectly names QtWebEngineProcess
due to "misplaced" Info.plist file from the bundle.