hooks: remove run-time hooks for QtWebEngine
We do not need to override QTWEBENGINEPROCESS_PATH anymore, as
the helper executable's location is now correctly inferred from
the shared library location within the .framework bundle.

And with PySide6/PyQt6 we do not need to disable sandboxing anymore,
either - it should work with the properly preserved
QtWebEngineCore.framework bundle layout.