building: post-process binaries TOC with work-arounds for pywin32
At the end of the analysis process, post-process the `binaries`
TOC and apply two work-arounds for the `pywin32` package on
Windows.

First, we now attempt to preserve the parent directory layout for
`pywin32` extensions that are originally located in `win32` and
`pythonwin` directories. The `pywin32.pth` places these directories
into `sys.path`, so we end up seeing them as top-level modules,
and collect them into `sys._MEIPASS`. However, some of these
extensions might be linked to other extensions, and for those our
binary dependency analysis would preserve the parent directory
layout. So to keep things consistent, we now explicitly attempt
to preserve the parent directory of extensons that were discovered
by the modulegraph as well. This is done as post-processing step
on the `binaries` TOC, and also requires us to add the
sub-directories in question to `sys.path` in our bootstrap module
for `pywin32`.

The second work-around is specific to Anaconda `pywin32` package,
which installs three copies of `pywintypes3X.dll` and `
pythoncom3X.dll`, in the following locations:
- Library/bin
- Lib/site-packages/pywin32_system32
- Lib/site-packages/win32

So if, due to order of imports and library location search order,
we end up collecting from `Library/bin` (into `sys._MEIPASS`) or
from `Lib/site-packages/win32` (into `sys._MEIPASS/win32`), we need
to explicitly change the destination directory to
`sys._MEIPASS/pywin32_system32` to match the layout we expect.