depend: bindepend: remove special handling for pywin32
Remove the special handling for `pywin32` from dependency scanner's
parent directory preservation mechanism. We had code in place that
explicitly prevented directory preservation for dependent binaries
from `pythonwin` directory.

This was because we were collecting `win32uiole` extension from that
directory into `sys._MEIPASS` (due to it being seen as top-level
module, thanks to `pywin32.pth` adding the directory to `sys.path`);
but the extension is linked against `win32ui` extension, so when
discovered by dependency analysis of `win32uiole`, we would place it
in the `pythonwin` directory. Thus, the two extensions would end
up in different place, and `win32uiole` would refuse to load due to
missing dependency.

But now, our analysis/build process explicitly forces collection
of extensions from `pythonwin` and `win32` to their original
subdirectories (by means of `binaries` TOC post-processing), so
it does not matter if we collect `win32ui` into the subdirectory;
the `win32uiole` will also be placed next to it in the end.