bindepend: further try to preserve .framework bundles
When we are collecting a shared library into top-level directory
and this shared library comes from a .framework bundle, re-create
the .framework bundle directory structure in the top-level
directory, and create symlink to library for backward compatibility.

The Info.plist should then also be automatically collected as
part of previously-added collection mechanism.

This functionality aims to further preserve .framework bundle
structures, this time in cases when they are collected from
system-wide library directory instead of from a package; for
example, when collecting Qt .framework bundles from Homebrew
installation.