loader: PyiFrozenImporter: work-around for importlib's LazyLoader
When `importlib.util.LazyLoader` is used, the `module` passed to
the `PyiFrozenImporter.exec_module` contains modified spec, in
which the `loader_state` is a `dict` prepared by the `LazyLoader`
code.

This means that our original `loader_state` is lost, and we need
to determine the module name for look-up in the PYZ by other means.
Specifically, we convert the `origin` back to the module name, which
allows us to handle aliased modules as well.