building: macOS: use @rpath to rewrite binaries' dependency paths
When rewriting the dylib identifier and paths to linked
non-system shared libraries in collected binaries, use
`@rpath` instead of directly using `@loader_path`.

E.g., for a binary collected as `mypackage/lib/libsomething.dylib`,
instead of rewriting the identifier and paths to linked dylibs as
`@loader_lib/../../<something>`, rewrite as `@rpath/<something>`
and set rpath to `@loader_path/../..`.

This ensures that the dylib identifier of a shared library
(now `@rpath/libsomething.dylib`) always matches the name/path
used by another collected binary to refer to this shared library
(the dependency path now also being `@rpath/libsomething.dylib`),
even if the shared library and referring binary are located at
different directory depth w.r.t. the top-level application
directory. Under the old approach that directly uses
`@loader_path`, the dylib identifier of shared library might
end up being `@loader_path/../../libsomething.dylib`, but the
library path in the binary located one directory higher would
end up being `@loader_path/../libsomething.dylib`.

Mismatches in dylib identifiers are typically not a problem,
because the linked dylib is actually found at the specified
location. However, some packages do not place shared libraries
in "correct" location, and work around this by "preloading"
the shared libraries from another location, for example via
`ctypes`. In this case, for a binary to recognize an
already-loaded library, the referring name and the target
dylib's identifier must match. And in PyInstaller's case, the
only way to achieve that is to have identifiers (and referring
paths) normalized to `@rpath/<name>`, and store the actual
relative path to the top-level application directory in the
rpath.

Using `@rpath`, however, means that we now need to ensure that
our rpath is the only one set in the binary; i.e., we need to
remove all other rpaths, and add our target rpath if necessary.
To do so, we need to use the external macOS-provided
`install_name_tool` utility, which we now use for all header
modifications: modification of dylib identifier, modification
of dependend dylib paths, and removal/addition of rpaths.