building: BUNDLE: rewrite the resource relocation mechanism
Redesign and rewrite the resource relocation mechanism used to divert
data files into `Contents/Resources` directory while keeping the
binaries in `Contents/MacOS`.

The relocation is now performed as a pre-processing step on the input
TOC, and produces the "final" TOC, which is then used by the `assemble`
method to preform the actual collection. The separation of relocation
step from the collection step makes it easier to make adjustments to
the relocation implementation. At the same time, it significantly
simplifies the collection itself, with the code there now being very
similar to what we have in the COLLECT target.

The relocation mechanism has been completely redesigned. We now
relocate ALL data files to `Contents/Resources`. That includes the
source  .py and byte-compiled .pyc files (which were exempted from
relocation in # in an attempt to fix `cv2` loader scripts expecting
to find the extension binary next to source .py file). It also includes
data files from the PySide2/PySide6/PyQt5/PyQt6 which have been
explicitly excluded from relocation up until now.

The nested .framework bundles are treated as monolihic BINARY
entities, and thus kept in `Contents/MacOS` (i.e., the DATA entries
from within such .framework bundles are not relocated).

While previously only data files (and data-only directories) have
been symlinked from `Contents/Resources` back to `Contents/MacOS`,
we now automatically cross-link all files between the two directory
structures in an attempt to maintain the illusion of mixed-content
directories at either location. Data-only and binary-only directories
are cross-linked at directory level, while files in mixed-content
directories are cross-linked at individual file level.

An automatic work-around for directories in `Contents/MacOS` that
have dot in their name is implemented; the directory is created
with a modified name, and a symbolic link pointing to it is created
under the original name.

A series of tests has been added to verify that the basic rules of
relocation.