building: fix preserved .framework bundle structure
When collecting the `Info.plist file`, we need to collect the
original file from `Versions/<version>/Resources`, rather than
from symlinked `Resources` in the top-level directory. Otherwise
`codesign` verification reports `embedded framework contains
modified or invalid version`.

We also need to (re)create the symlink `Versions/Current` pointing
to `Versions/<version>` directory from which we collected the
binary. Framework bundles shipped with python packages may not
include this, but it is required by `codesign`, so we always
manually create this symlink (i.e., do not attempt to collect it
even if it is present).

Neither signing nor verification with `codesign` seem to require
us to symlink `Versions/Current/Resources` and `Versions/Current/<binary>`
to the top-level directory, so for now, we do not do that.