building: fix preserved .framework bundle structure, try 2
While the well-formed versioned .framework bundles should contain
their `Info.plist` in the `Versions/<version>/Resources` directory,
we are likely to come across the bundles that contain `Info.plist`
in the top-level `Resources` directory (while the binary is in the
versioned directory). This seems to be the case with contemporary
`PySide2`, `PyQt5`, and `PyQt6`.

Accommodate such cases, but ensure that the file is collected into
versioned `Resource` directory, which seems to be a requirement
for code-signing.