building: fix preserved .framework bundle structure, try 3
As part of post-processing of collected files from .framework bundles,
re-create the symlinks to the content within the `Versions/Current`
(which itself is a re-created symlink to the `Versions/<version>`)
from the top-level .framework directory: the binary file, the
Resources directory, and auxiliary directories, such as Helpers.

While this is a generic solution, it is primarily aimed at
seamlessly accommodating the QtWebEngine in contemporary PySide/PyQt,
where the QtWebEngineProcess helper process is sought in the
top-level Helpers directory and the resources are sought in the
top-level Resources directory. On the other hand, we need to collect
these directories in the `Versions/<version>` directory to be
compliant with codesign...