hookutils: qt: fix collection of extras from QtWebEngineCore.framework
The Resources and Helpers directory that we collect from the
QtWebEngineCore.framework need to be collected in the versioned
directory (Versions/<version>/Helpers and Versions/<version>/Resources)
in order to be compliant with codesign, so adjust the paths
accordingly.

The .framework bundles shipped with contemporary PyPI PyQt/PySide wheels
are naturally not well-formed, so we need to fall-back to collecting
stuff from top-level .framework directory. But collect into versioned
directory to be compliant with codesign.

The helper process and resources are actually sought in the top-level
.framework directory (which is why everything except the codesign
worked so far) - this is now handled by symlinks to the top-level
directory that are automatically created by .framework post-processing
code in the `PyInstaller.utils.osx.collect_files_from_framework_bundles`
helper function (see preceding commit).