rthooks: pyqt/pyside: extend qml import paths in macOS .app bundles
To satisfy codesign requirements, we are forced to split the
collected `qml` directory into two parts; one that keeps only
binaries (rooted in the `Contents/MacOS`) and one that keeps
only data files (rooted in the `Contents/Resources`), with files
from one directory tree being symlinked to the other to mantinain
illusion of a single mixed-content directory.

As Qt seems to compute the identifier of its QML components based
on location of the `qmldir` file w.r.t. the registered QML import
paths, we need to register both paths, because the `qmldir` file
for a component could be reached via either directory tree.