tests: splash: extend splash tests with tkinter-enabled variant
Extend the existing splash tests (onedir, onefile) that do not
collect tkinter (and, in fact, explicitly exclude it) with a
variant that forces tkinter collection via hiddenimports.

The `tcl86t.dll` shipped with Windows build of Python 3.12b1
depends on `zlib1.dll`, and due to the lack of dependency scanning
in splash, we fail to pick it up there. Threrefore, only
`tkinter-onedir` test variant succeeds, as the `zlib1.dll`
dependency is picked up by dependency scanning in main build
process, which analyzes Tcl/Tk DLLs as dependencies of `tkinter`.

Both `notkinter` test variants fail to collect `zlib1.dll`, while
`tkinter-onefile` collects it, but fails to extract it as it is
not listed among the splash screen requirements.