splash: perform binary dependency analysis on Tcl/Tk shared libraries
Perform binary dependency analysis on Tcl/Tk shared libraries, to
ensure that a) they are collected in cases when the progam does
not use `tkinter`, and b) they are made available in onefile
builds via pre-extraction.

Up until now, the only dependency we had to worry about was
`vcruntime140.dll` on Windows (and even that only for onefile
pre-extraction), but the `tcl86t.dll` that comes with Windows
build of Python 3.12b1 also depends on `zlib1.dll`.