test/bytecode: Update test for scanning nested code objects.
In Python <= 3.11, the compiled bytecode for [foo() for i in range(10)]
is two code objects - the code object for the `foo()` bit being embedded
in the code object for everything else. In Python >= 3.12, it's one flat
code object. This makes it unsuitable as a test for scanning nested code
objects since it's only the correct situation on older Pythons.

Swap out the function call in the body of a comprehension loop with
calling a function in a lambda which still does the right thing with any
Python version.