building: BUNDLE: relocate dylibs to Contents/Frameworks directory
When assembling macOS .app bundle, put the dylibs and nested
.framework bundles into `Contents/Frameworks` directory, so that
only the program executable is left in the `Contents/MacOS`.

This requires us to relocate `sys._MEIPASS` from `Contents/MacOS`
(= parent of the executable directory) to `Contents/Frameworks`
in lieu of having to cross-link everything back (which would
defeat the purpose of the relocation in the first place).

A side effect is that `sys._MEIPASS` does not correspond to
`os.path.dirname(sys.executable)` in onedir .app bundles anymore.
But that never held for onefile builds (regular or .app bundles),
and may not hold for regular onedir builds in the near future.