hooks: add pre-safe-import-module hook for distutils
Starting with python 3.12, `distutils` is not part of stdlib anymore,
and it is provided solely by `setuptools`. Due to limitations of the
modulegraph, the latter cannot find and properly analyze the
`setuptools`-provided `distutils` (and its submodules) under mapped
names, so we need the pre-safe-import-module hook to mark the
`distutils` as a run-time package.

Marking the package as a run-time one means that we also need to
mark its submodules and subpackages as run-time ones.