config: clean up UPX availability check
Do not attempt to parse UPX version string, just display it for
information. Also switch from `compat.exec_command` to plain
`subprocess.check_output`.

The only reason we were trying to parse the version was
a) to catch an ancient versions of UPX (< 1.92) that were
apparently incompatible with python 2.4, and
b) to use the --lzma switch only if UPX 3.00 or newer is
available.

UPX 3.00 was released in 2007, so I think it is safe to remove
the version checks.