building: clean up UPX handling
When disabling UPX due to platform-specific issues, do so during
initial setup instead of right before we are about to use it.
Display INFO message if UPX is available and will be used if
enabled (on Windows) or if it is available but is disabled on
the platform (on non-Windows).

Stop setting default UPX flags via the `UPX` environment variable;
instead, set those flags explictly on command-line.

Rename the `hasUPX` `CONF` variable into `upx_available`.