Make the name of onedir mode's `_internal` directory configurable.
Add a --contents-directory CLI option and corresponding
contents_directory spec file option to change the otherwise hard-coded
name of the directory in which all a onedir's non EXE contents are
hidden.

Note that this, whilst this spec file option would make the most sense
as a parameter to BUNDLE(), it instead has to be an option to EXE()
which BUNDLE() then fishes out of the EXE()'s configuration due to EXE()
being the only place where OPTION TOC entries can go.