bulding: BUNDLE: avoid cross-linking the EXECUTABLE file
The .app bundle's executable's location should be always
discoverable via `sys.executable`, so there is no reason to
cross-link it into `Contents/Frameworks` and `Contents/Resources`.

By not cross-liking the executable, we also prevent potential
conflicts between the executable (i.e., program/app bundle name)
and an eponymous package (from which we collect binaries or data
files), which arise in scenario when same base name is used for
the entry-point script (`something.py`) and a package (`something`).