Add option to hide/minimize console in console-enabled Windows program
Add new EXE argument (`hide_console`) and command-line option
(`--hide-console`) with options `hide-early`, `hide-late`,
`minimize-early`, and `minimize-late`; these are mapped to a
bootloader option that instructs the bootloader to programmatically
hide or minimize the console window, either as soon as possible
or as late as possible. The operation is performed only if the
process owns the console window (i.e., the program was not launched
from an existing console window).

Applies only to console-enabled builds on Windows.