building: improve handling of symlink chains with uncollected links
If we have a symbolic link chain with intermediate links that we
do not collect, attempt to rewrite the links to "jump over" the
missing ones.

This attempts to mitigate potential issues under Homebrew python,
where shared libraries for `wxwidgets` have the following layout:
 * libwx_baseu-3.2.0.2.1.dylib
 * libwx_baseu-3.2.0.dylib -> libwx_baseu-3.2.0.2.1.dylib
 * libwx_baseu-3.2.dylib -> libwx_baseu-3.2.0.dylib

If the other collected binaries reference only the two of them,
`libwx_baseu-3.2.0.2.1.dylib` and `libwx_baseu-3.2.dylib`, we
end up missing the intermediate link, and the final link ends
up as a hard copy, causing issues due to duplication.

Therefore, if we do not seem to be collecting the file referenced
by a symlink, but the referenced file itself is a symlink, follow
it again and see if we happen to be collecting that file (and if
that file happens to be a symlink, follow it again...).