building: rework symbolic link check to handle linked parent dir
Rework the symbolic link check to handle the problem with linked
parent directory, as observed on macOS:
 * /usr/local/Cellar/wxwidgets/3.2.2.1_1/lib/libwx_baseu-3.2.0.2.1.dylib
 * /usr/local/Cellar/wxwidgets/3.2.2.1_1/lib/libwx_baseu-3.2.0.dylib -> libwx_baseu-3.2.0.2.1.dylib
 * /usr/local/Cellar/wxwidgets/3.2.2.1_1/lib/libwx_baseu-3.2.dylib -> libwx_baseu-3.2.0.dylib
and
 * /usr/local/opt/wxwidgets/lib/libwx_baseu-3.2.0.2.1.dylib
 * /usr/local/opt/wxwidgets/lib/libwx_baseu-3.2.0.dylib -> libwx_baseu-3.2.0.2.1.dylib
 * /usr/local/opt/wxwidgets/lib/libwx_baseu-3.2.dylib -> libwx_baseu-3.2.0.dylib
which are actually the same, because
 * /usr/local/opt/wxwidgets -> ../Cellar/wxwidgets/3.2.2.1_1

Other binaries end up referencing
`/usr/local/opt/wxwidgets/lib/libwx_baseu-3.2.dylib` and
`/usr/local/Cellar/wxwidgets/3.2.2.1_1/lib/libwx_baseu-3.2.0.2.1.dylib`.

So in addition the the problem with chained links and missing
intermediate link, we also need to be able to handle the situation
where the files appear to be originating from different locations
due to the linking of parent directories.