Add Nvidia libraries to Unix exclusion list (#7746)
While packaging an application that uses pynvml on Linux, the library 
libnvidia-ml.so.1 may be bundled into the resulting application. This 
library is part of the Nvidia driver stack and this behavior creates 
an incompatibility if the packaging system and runtime system do not 
have the same driver version (or are compatible in some vague sense I 
don't understand).

For my application and I assume pretty much every application that 
might want to bundle pynvml this way, libnvidia-ml.so.1 should be 
excluded from the final distribution.