building: remove support for external executable manifest
Remove support for generating external (side-by-side) manifest
for the Windows executable. This has been available for onedir
builds, but is now broken due to relocation of all bundle contents
into sub-directory. Instead of trying to fix it, simply drop the
option; if anyone needs to post-process the manifest, they can
used 3rd party tools to extract and re-embed it.