building: remove support for WinSxS assemblies analysis and collection
Remove support for analysis and collection of WinSxS assemblies.
The amount of code involved is disproportionate to the chance of
encountering a binary making use of WinSxS (and the required
assembly not being available on the target system) in contemporary
python and python packages (e.g., VC9 runtime from VS2008 was the
last VC runtime that made use of WinSxS; and was used to build
python 2.7 and python <= 3.3).