building: remove PyInstaller.building.toc_conversion module
Remove the `building.toc_conversion` module and its
`DependencyProcessor` class. The main purpose of this class was
aggregation of `binaries` and `datas` TOC lists returned from
the hooks, with additional collection of files from eggs.

In contemporary python environments, however, that extra data
collection should be handled by hooks, which should work with
non-zipped eggs as well as wheel-installed packages. So remove
the `DependencyProcessor` and replace it with two helper methods
on `PyiModuleGraph` that return `binaries` and `datas` TOC lists
for entries returned by hooks.

This removes the support for zipped eggs from PyInstaller;
the `zipfiles` and `zipped_data` TOC lists of Analysis should now
be always empty.