tests: adjust/update for removal of zipped egg support
In binary-vs-data-reclassification tests, adjust the indices of
`binaries` and `datas` TOCs in the serialized `Analysis` data
struct; this is needed because elements were removed from the
object's guts.

Fix `test_egg_unzipped`. The test expects the data file from
egg to be automatically collected, but it should in fact be
collected via hook. So add the said hook.

Remove the `test_egg_zipped`, as zipped eggs are not supported
anymore.

Remove the egg-based variant of tests for `pkg_resources` provider.
As we do not support zipped eggs anymore, drop the (zipped) egg
variant of pkg_resources provider tests.