winutils: simplify winmanifest module
Replace the manifest parser (and processor) in winmanifest module
with a simplified version.

The former parser/processor included code for WinSxS assembly
processing, look-up and binding redirection; with support for
WinSxS processing dropped from PyInstaller, this is not needed
anymore.

If provided with manifest XML, the former parser attempted to
load it into its structures, and then write them back into XML
when embedding the manifest into the executable. Due to processing
only a subset of manifest XML schema, this inherently meant that
elements unknown to the parser were lost in the process. Which in
turn negated main reason for providing external manifest (aside
from being able to set custom application name and version).

The new manifest processing code encodes the default manifest as
an XML, making it easier to read and edit. When creating application
manifest, the base XML (the template or externally-provided XML)
is processed just to:

* write uac_admin and uac_uiaccess options into corresponding
  attributes of the <requestedExecutionLevel> element (creating it
  and its parents if necessary)

* ensure that <dependency> element exists for MS Common-Controls
  dependent assembly, and that the attributes of corresponding
  <assemblyIdentity> element are set to the expected values
  (this is required by windowed bootloader for the unhandled
  exception dialog).

All other parts of the manifest XML are left untouched.

The new manifest processing does not require writing manifest
to the disk any more, as all processing is done in-memory.

The winmanifest module can now be imported on all platforms, as
`winresource` module is imported only within the functions that
read and write the manifest from/to the executable.