winutils: icon: have icon group IDs start at 1 instead of 0
Having icon group with ID 0 seems to abort the enumeration of
resources using `winresource.GetResources` helper with
`win32ctypes.pywin32.pywintypes.error: (15106, 'EnumResourceNamesW',
'User stopped resource enumeration.')`. Presumably this is because
the 0 ID evaluates to falsely value somewhere in the underyling
win32-ctypes code.

Even if this is underlying win32-ctypes problem, Windows executables
seem to have their icon group IDs start with 1, so keep in line
with that.