bindepend: improve suppression of missing-lib warnings
As the missing-library warnings are now handled by the high-level
binary dependency analysis function, we can be a bit smarter about
them, and suppress the warnings for dependencies that happen to
be collected anyway.

This might happen either because the library was collected by some
other mechanism (for example, via hook, or supplied by the user),
or because it was discovered during the analysis of another binary
(which, for example, had properly set run-paths on Linux/macOS or
was located next to that other analyzed binary on Windows).

Therefore, we defer the missing-library warnings until after binary
dependency analysis is complete; at that point, we can compare the
basenames of unresolved dependencies to basenames of discovered
(collected) binaries for improved warning suppression.