building: find_binary_dependencies: re-enable retrieval of PATH modifications
Re-enable retrieval of PATH modifications from the isolated
subprocess in which `find_binary_dependencies` is importing collected
packages. In contrast to original implementation, retrieve only
changes made to PATH instead of complete PATH. This makes the
corresponding log message easier to read, and also prevents redundant
search path entries, as paths from the original PATH are automatically
added in the underyling library-search function.

Remove bunch of comments that are obsoleted by the fact that
`find_binary_dependency` itself is not ran in the isolated
process itself.