hookutils: gi: fix error handling in compile_glib_schema_files
Fix error handling in compile_glib_schema_files; do not attempt
to access attributes of `p`, which is unavailable in the exception
handling block. To display stdout/stderr of the failed process,
use attributes of the `subprocess.CalledProcessError` instead.

Ignore the character encoding errors when decoding stdout/stderr
from the `glib-compile-schema` process. The warnings it emits
seem to use `U+201C` and `U+201D` quotation marks, which might
cause issues with active locale (for example, in msys2/mingw64
environment).