bootloader: move the file pointer out of ARCHIVE_STATUS structure
The archive file access in pyi_arch_* functions is always
encapsulated to the function; i.e., the function opens the file,
reads the contents, and then closes the file (except for faulty
error handling in pyi_arch_open, which is fixed now).

 Therefore, it makes no sense to keep the track of file pointer in
the ARCHIVE_STATUS.

Moving it out cleans up the function signatures; the ARCHIVE_STATUS
argument can be constified, and low-level helpers can be passed
the file pointer instead of the ARCHIVE_STATUS struct.