bootloader: dump exception info if Py_InitializeFromConfig fails
In case we fail to initialize the python interpreter using
`Py_InitializeFromConfig`, use ˙Py_ExitStatusException˙ to dump
the exception information to stderr.

It also immediately exits the  process with error code, which
somewhat clashes with our error handling design; but the benefit
of getting the extra diagnostic information (when console is
enabled) trumps over that.