building: EXE: extend the operation retry mechanism
Extend the operation retry mechanism to cover the following
exception types:
 - PermissionError
 - OSError with errno 13 or 22
 - pywintypes.error with winerror 5 and 32

The retry mechanism is now applied to all processing steps that
involve modifying the generated executable in Windows-specific
codepaths of `EXE.assemble` as well as shared codepaths that are
used on Windows.

The retry message has been extended, and promoted from DEBUG to
WARNING to make it obvious that retrying is going on in the build
environment.